rework file structure for store modules and setup methods
