+++ src/store/features/bookmarks.slice.ts
@@ -1,34 +1,34 @@
-import { type AsyncThunkPayloadCreator, createEntityAdapter, type EntityId, isAnyOf } from "@reduxjs/toolkit";
+import { type AsyncThunkPayloadCreator, asyncThunkCreator, buildCreateSlice, createEntityAdapter, type EntityId, type GetThunkAPI, isAnyOf } from "@reduxjs/toolkit";
 import { sortObjectFn } from "bsmap";
 
 import { resolveBookmarkId, resolveColorForBookmark } from "$/helpers/bookmarks.helpers";
 import { addSong, leaveEditor, loadBeatmapEntities, startLoadingMap } from "$/store/actions";
-import { createSlice } from "$/store/helpers";
 import { selectCursorPositionInBeats } from "$/store/selectors";
-import type { RootState } from "$/store/setup";
+import type { AppThunkApiConfig } from "$/store/types";
 import type { App, RequiredKeys, SongId, View } from "$/types";
 
 const adapter = createEntityAdapter<App.IBookmark, EntityId>({
 	selectId: resolveBookmarkId,
 	sortComparer: sortObjectFn,
 });
+
 const { selectAll } = adapter.getSelectors();
 
-const deriveDataFromState: AsyncThunkPayloadCreator<RequiredKeys<App.IBookmark, "time">, { songId: SongId; view: View; name: string }> = (args, api) => {
-	const state = api.getState() as RootState;
-	// we want to use the cursor position to figure out when to create the bookmark for
-	const beatNum = selectCursorPositionInBeats(state, args.songId);
-	if (beatNum === null) return api.rejectWithValue("Invalid beat number.");
-	return api.fulfillWithValue({ ...args, time: beatNum, color: resolveColorForBookmark(args.name) });
-};
-
-const slice = createSlice({
+const slice = buildCreateSlice({ creators: { asyncThunk: asyncThunkCreator } })({
 	name: "bookmarks",
 	initialState: adapter.getInitialState(),
 	selectors: {
 		selectAll: selectAll,
 	},
 	reducers: (api) => {
+		const deriveDataFromState: AsyncThunkPayloadCreator<RequiredKeys<App.IBookmark, "time">, { songId: SongId; view: View; name: string }> = (args, api: GetThunkAPI<AppThunkApiConfig>) => {
+			const state = api.getState();
+			// we want to use the cursor position to figure out when to create the bookmark for
+			const beatNum = selectCursorPositionInBeats(state, args.songId);
+			if (beatNum === null) return api.rejectWithValue("Invalid beat number.");
+			return api.fulfillWithValue({ ...args, time: beatNum, color: resolveColorForBookmark(args.name) });
+		};
+
 		return {
 			addOne: api.asyncThunk(deriveDataFromState, {
 				fulfilled: (state, action) => {
