+++ src/store/features/clipboard.slice.ts
@@ -1,9 +1,8 @@
-import type { AsyncThunkPayloadCreator, CaseReducer, PayloadAction } from "@reduxjs/toolkit";
+import { type AsyncThunkPayloadCreator, asyncThunkCreator, buildCreateSlice, type CaseReducer, type GetThunkAPI, type PayloadAction } from "@reduxjs/toolkit";
 import { sortObjectFn } from "bsmap";
 
-import { createSlice } from "$/store/helpers";
 import { selectSelectedBeatmapEntities } from "$/store/selectors";
-import type { RootState } from "$/store/setup";
+import type { AppThunkApiConfig } from "$/store/types";
 import type { App, View } from "$/types";
 
 const initialState = {
@@ -11,29 +10,7 @@ const initialState = {
 	data: {} as Partial<Omit<App.IBeatmapEntities, "bookmarks">>,
 };
 
-const fetchClipboardData: AsyncThunkPayloadCreator<typeof initialState, { view: View }> = (args: { view: View }, api) => {
-	const state = api.getState() as RootState;
-	const selection = selectSelectedBeatmapEntities(state, args.view);
-	return api.fulfillWithValue({ ...args, data: selection });
-};
-
-const processSelection: CaseReducer<typeof initialState, PayloadAction<typeof initialState>> = (state, action) => {
-	const { data } = action.payload;
-	if (!data) return state;
-	return {
-		...state,
-		data: {
-			// We want to sort the data so that it goes from earliest beat to latest beat.
-			notes: data.notes?.sort(sortObjectFn),
-			bombs: data.bombs?.sort(sortObjectFn),
-			obstacles: data.obstacles?.sort(sortObjectFn),
-			basicEvents: data.basicEvents?.sort(sortObjectFn),
-			boostEvents: data.boostEvents?.sort(sortObjectFn),
-		},
-	};
-};
-
-const slice = createSlice({
+const slice = buildCreateSlice({ creators: { asyncThunk: asyncThunkCreator } })({
 	name: "clipboard",
 	initialState: initialState,
 	selectors: {
@@ -52,6 +29,27 @@ const slice = createSlice({
 		},
 	},
 	reducers: (api) => {
+		const fetchClipboardData: AsyncThunkPayloadCreator<typeof initialState, { view: View }> = (args: { view: View }, api: GetThunkAPI<AppThunkApiConfig>) => {
+			const state = api.getState();
+			const selection = selectSelectedBeatmapEntities(state, args.view);
+			return api.fulfillWithValue({ ...args, data: selection });
+		};
+		const processSelection: CaseReducer<typeof initialState, PayloadAction<typeof initialState>> = (state, action) => {
+			const { data } = action.payload;
+			if (!data) return state;
+			return {
+				...state,
+				data: {
+					// We want to sort the data so that it goes from earliest beat to latest beat.
+					notes: data.notes?.sort(sortObjectFn),
+					bombs: data.bombs?.sort(sortObjectFn),
+					obstacles: data.obstacles?.sort(sortObjectFn),
+					basicEvents: data.basicEvents?.sort(sortObjectFn),
+					boostEvents: data.boostEvents?.sort(sortObjectFn),
+				},
+			};
+		};
+
 		return {
 			cutSelection: api.asyncThunk(fetchClipboardData, {
 				fulfilled: processSelection,
