+++ src/store/hooks.ts
@@ -1,6 +1,6 @@
 import { useDispatch, useSelector } from "react-redux";
 
-import type { AppDispatch, RootState } from "./setup";
+import type { AppDispatch, RootState } from "./types";
 
 export const useAppSelector = useSelector.withTypes<RootState>();
 export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
