+++ src/store/middleware/backup.middleware.ts
@@ -3,7 +3,7 @@ import { createListenerMiddleware, isAnyOf, type PayloadAction } from "@reduxjs/
 import { createSaveHandler } from "$/services/backup.service";
 import { leaveEditor, saveBeatmapContents, updateBeatmap, updateSong } from "$/store/actions";
 import { selectSelectedBeatmap } from "$/store/selectors";
-import type { AppDispatch, RootState } from "$/store/setup";
+import type { AppDispatch, RootState } from "$/store/types";
 import type { App, BeatmapId, SongId } from "$/types";
 
 export default function createBackupMiddleware() {
